+++ spec/spec_helper.rb
@@ -123,6 +123,13 @@ RSpec.configure do |config|
       llm_message_inference_requests
       llm_messages
       llm_conversations
+      portable_identity_group_memberships
+      portable_identity_audit_log
+      portable_identities
+      portable_identity_groups
+      portable_identity_principals
+      portable_identity_provider_capabilities
+      portable_identity_providers
     ].each do |table|
       Legion::Data.connection[table].delete if Legion::Data.connection.table_exists?(table)
     end